gops-clients/python/pyproject.toml
[tool.poetry]
name = "py-gops-players"
version = "1.0.0"
description = "Examples for GOPS client in Python"
authors = ["Matthew <matthewshin2@gmail.com>"]
readme = "README.md"
packages = [{include = "py_gops_players"}]

[tool.poetry.dependencies]
python = "^3.10"
websocket-client = "^1.5.1"
rel = "^0.4.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
